+++ Editor/Scripts/Typed/StringIntGameEventInspector.cs
@@ -1,0 +1,9 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(StringIntGameEvent))]
+  public class StringIntGameEventInspector : TwoTypeGameEventInspector<string, int>
+  {
+  }
+}