+++ Runtime/Scripts/TypedGameEvent/Generic/TwoTypeGameEvent.cs
@@ -1,3 +1,5 @@
+using UnityEngine.Events;
+
 namespace FredericRP.EventManagement
 {
   public class TwoTypeGameEvent<T, U> : GameEvent
@@ -18,6 +20,15 @@ namespace FredericRP.EventManagement
       this.firstParameter.value = parameter;
       this.secondParameter.value = secondParameter;
       Raise(eventHandler);
+    }
+
+    public void Listen(UnityAction<T, U> action, GameEventHandler eventHandler = null)
+    {
+      base.Listen<T, U>(action, eventHandler);
+    }
+    public void Delete(UnityAction<T, U> action, GameEventHandler eventHandler = null)
+    {
+      base.Delete<T, U>(action, eventHandler);
     }
   }
 }