+++ Runtime/Scripts/TypedGameEvent/Typed/StringIntGameEvent.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/String-Int Game Event")]
+  public class StringIntGameEvent : TwoTypeGameEvent<string, int>
+  {
+  }
+}